Fix of missing ZipArchive package

+++ partner/index.php
@@ -5,25 +5,27 @@ if(isset($REQ["update"]) && $REQ["update"] == "true"):
     die();
 endif;
 if(!isset($IPP_CONFIG["INTERACTIVE_GUIDE"])) {
-    $src = BASEDIR."plugins/interactive_guide/";
-    $filename = $src . basename("https://plugins.ippworldwide.com/interactive_guide.zip");
-    $dirMode = 0755;
-    if(!file_exists($src))
-        if (!mkdir($src, $dirMode, true) && !is_dir($src)) {
-            throw new \RuntimeException(sprintf('Directory "%s" was not created', $src));
+    if(class_exists('ZipArchive')) {
+        $src = BASEDIR."plugins/interactive_guide/";
+        $filename = $src . basename("https://plugins.ippworldwide.com/interactive_guide.zip");
+        $dirMode = 0755;
+        if(!file_exists($src))
+            if (!mkdir($src, $dirMode, true) && !is_dir($src)) {
+                throw new \RuntimeException(sprintf('Directory "%s" was not created', $src));
+            }
+        sleep(1);
+        file_put_contents($filename, fopen("https://plugins.ippworldwide.com/interactive_guide.zip", 'r'));
+        $zip = new ZipArchive();
+        $res = $zip->open($filename);
+        if ($res === TRUE) {
+            $zip->extractTo($src);
+            $zip->close();
+            $partner->InstallPlugin("interactive_guide");
+        } else {
+            throw new \RuntimeException(sprintf('Could not Unzip file at "%s"', $src));
         }
-    sleep(1);
-    file_put_contents($filename, fopen("https://plugins.ippworldwide.com/interactive_guide.zip", 'r'));
-    $zip = new ZipArchive();
-    $res = $zip->open($filename);
-    if ($res === TRUE) {
-        $zip->extractTo($src);
-        $zip->close();
-        $partner->InstallPlugin("interactive_guide");
-    } else {
-        throw new \RuntimeException(sprintf('Could not Unzip file at "%s"', $src));
+        unlink($filename);
     }
-    unlink($filename);
     include(BASEDIR . "controller/IPPConfig.php");
     $config = new IPPConfig();
     $config->UpdateConfig("INTERACTIVE_GUIDE","1");
